backend/requirements.txt
aiohttp==3.10.*
aiomysql==0.2.0
aioresponses==0.7.6
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
anytree==2.12.1
APScheduler==3.10.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
audioread==3.0.1
av==12.2.0
awscli==1.29.54
beautifulsoup4==4.12.3
camel-converter==3.1.1
cdifflib==1.2.6
certifi==2023.7.22
channels==4.0.0
chardet==5.2.0
charset-normalizer==3.3.0
cleo==2.1.0
click==8.1.7
comm==0.2.1
contourpy==1.2.1
crashtest==0.4.1
cycler==0.12.1
Cython==3.0.0
debugpy==1.8.1
decorator==4.4.2
distro==1.9.0
Django==5.0.2
docutils==0.16
dulwich==0.21.7
editdistance==0.8.1
einx==0.3.0
encodec==0.1.1
exceptiongroup==1.1.3
executing==2.0.1
fastapi==0.103.2
fastapi-cli==0.0.4
fastjsonschema==2.19.1
filelock==3.12.4
frozendict==2.4.4
frozenlist==1.4.0
fsspec==2023.9.2
ghc @ https://raw.githubusercontent.com/AwaleSajil/ghc/master/ghc-1.0-py3-none-any.whl
gitdb==4.0.11
gitignore-parser==0.1.9
gobject==0.1.0
GPUtil==1.4.0
h11==0.14.0
html2docx==1.6.0
httpcore==1.0.5
httpx==0.27.2
idna==3.4
imageio==2.33.1
imageio-ffmpeg==0.4.9
imap-tools==1.5.0
iniconfig==2.0.0
installer==0.7.0
ipykernel==6.29.4
ipython==8.25.0
ipywidgets==8.1.1
jaraco.classes==3.4.0
jedi==0.19.1
Jinja2==3.1.2
jiter==0.5.0
joblib==1.3.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema-path==0.3.2
jsonschema-spec==0.2.4
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
kornia_rs==0.1.2
langchain==0.3.0
langchain-core==0.3.0
langchain-openai==0.2.0
langchain-text-splitters==0.3.0
langsmith==0.1.120
lazy-object-proxy==1.10.0
lazy_loader==0.4
libcst==1.0.1
lit==17.0.1
logger==1.4
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mdurl==0.1.2
meilisearch==0.21.0
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.4
nbclient==0.9.0
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.1
numpy==1.26.0
omegaconf==2.3.0
openai==1.45.0
openapi-schema-validator==0.6.2
openapi-spec-validator==0.7.1
opencv-python==4.10.0.84
orjson==3.10.7
packaging==24.1
parse==1.20.1
parso==0.8.3
pathable==0.4.3
pendulum==3.0.0
piexif==1.1.3
pkginfo==1.10.0
platformdirs==4.2.2
playwright==1.41.2
pluggy==1.5.0
pooch==1.8.1
prance==23.6.21.0
prettytable==3.10.0
proglog==0.1.10
prompt_toolkit==3.0.46
PuLP==2.8.0
pure-eval==0.2.2
py==1.11.0
pybase16384==0.3.7
pybase64==1.3.2
pycairo==1.26.1
pydantic==2.9.1
pydantic_core==2.23.3
pydub==0.25.1
pyee==11.0.1
Pygments==2.18.0
PyMuPDF==1.23.5
PyMuPDFb==1.23.5
PyMySQL==1.1.1
PyPDF2==3.0.1
pytest==8.1.1
pytest-asyncio==0.21.1
python-dateutil==2.9.0.post0
python-docx==0.8.11
python-multipart==0.0.17
PyYAML==6.0.1
pyzmq==25.1.2
ratelimiter==1.2.0.post0
regex==2023.8.8
requests==2.31.0
requests-toolbelt==1.0.0
retry==0.9.2
rfc3339-validator==0.1.4
rich==13.7.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
ruff==0.5.0
semantic-version==2.10.0
semver==3.0.2
shellingham==1.5.4
skypilot==0.4.1
smmap==5.0.1
sniffio==1.3.0
socksio==1.0.0
soundfile==0.12.1
soupsieve==2.6
soxr==0.3.7
SQLAlchemy==2.0.34
sqlparse==0.4.4
sse-starlette==2.1.3
stack-data==0.6.3
starlette==0.27.0
supervision==0.20.0
sympy==1.12
tenacity==8.5.0
threadpoolctl==3.2.0
tifffile==2024.4.24
tiktoken==0.7.0
time-machine==2.14.1
tinycss2==1.3.0
tomlkit==0.12.0
toolz==0.12.1
tornado==6.4
tqdm==4.66.5
traitlets==5.14.1
trampoline==0.1.2
trove-classifiers==2024.4.10
typeguard==4.3.0
typing_extensions==4.12.2
ujson==5.10.0
urllib3==1.26.16
uvicorn==0.23.2
uvloop==0.19.0
vector-quantize-pytorch==1.14.45
vocos==0.1.0
watchfiles==0.22.0
wcwidth==0.2.13
webencodings==0.5.1
wget==3.2
widgetsnbextension==4.0.10
wrapt==1.16.0
yarl==1.9.2
zhipuai==2.0.1
zxing-cpp==2.2.0
celery==5.3.4
redis==5.0.1
flower==2.0.1
